+++ htdocs/cabinetmed/core/modules/modCabinetMed.class.php
@@ -94,7 +94,7 @@ class modCabinetMed extends DolibarrModules
         $this->depends = array('modSociete');       // List of modules id that must be enabled if this module is enabled
         $this->requiredby = array();    // List of modules id to disable if this one is disabled
         $this->phpmin = array(4,3);                 // Minimum version of PHP required by module
-        $this->need_dolibarr_version = array(3,4,-3);   // Minimum version of Dolibarr required by module
+        $this->need_dolibarr_version = array(6, 0, -3);   // Minimum version of Dolibarr required by module
         $this->langfiles = array('cabinetmed@cabinetmed','companies');
 
         // Constants