+++ htdocs/cabinetmed/canvas/patient/tpl/card_view.tpl.php
@@ -252,13 +252,13 @@ if (empty($reshook) && ! empty($extrafields->attribute_label))
           	$agemonth=convertSecondToTime($now-$birthdate,'month')-1;
           	if ($ageyear >= 1) $newtmp.='('.$ageyear.' '.$langs->trans("DurationYears").')';
           	else if ($agemonth >= 1) $newtmp.='('.$agemonth.' '.$langs->trans("DurationMonths").')';
-          	else $newtmp.='('.$agemonth.' '.$langs->trans("DurationMonth").')';  	
+          	else $newtmp.='('.$agemonth.' '.$langs->trans("DurationMonth").')';
             //print $newtmp;
           	$tmp=preg_replace('/'.preg_quote('<td','/').'[^>]*'.preg_quote('colspan="3">'.dol_print_date($birthdate, 'day').'</td>','/').'/','<td colspan="3">'.dol_print_date($birthdate, 'day').$newtmp.'</td>',$tmp);
       	}
   	}
   	print $tmp;
-  	
+
   	print '<!-- end extrafields -->';
 }
 
@@ -425,7 +425,7 @@ if ($action == 'presend')
 	$formmail->fromid   = $user->id;
 	$formmail->fromname = $user->getFullName($langs);
 	$formmail->frommail = $user->email;
-    $formmail->trackid='pat'.$object->id;
+	$formmail->trackid='thi'.$object->id;
 	if (! empty($conf->global->MAIN_EMAIL_ADD_TRACK_ID) && ($conf->global->MAIN_EMAIL_ADD_TRACK_ID & 2))	// If bit 2 is set
 	{
 	    include DOL_DOCUMENT_ROOT.'/core/lib/functions2.lib.php';